+++ src/staticBrcode/staticBrcode.php
@@ -14,6 +14,7 @@ class StaticBrcode extends Resource
     public $keyId;
     public $city;
     public $amount;
+    public $cashierBankCode;
     public $reconciliationId;
     public $tags;
     public $uuid;
@@ -37,6 +38,7 @@ class StaticBrcode extends Resource
 
     ## Parameters (optional):
         - amount [integer, default 0]: positive integer that represents the amount in cents of the resulting Pix transaction. If the amount is zero, the sender can choose any amount in the moment of payment. ex: 1234 (= R$ 12.34)
+        - cashierBankCode [string, default ""] Cashier's bank code. ex: "20018183".
         - reconciliationId [string, default ""]: id to be used for conciliation of the resulting Pix transaction. ex: "123"
         - tags [array of strings, default []]: array of strings for tagging. ex: ["travel", "food"]
 
@@ -55,6 +57,7 @@ class StaticBrcode extends Resource
         $this-> keyId = Checks::checkParam($params, "keyId");
         $this-> city = Checks::checkParam($params, "city");
         $this-> amount = Checks::checkParam($params, "amount");
+        $this->cashierBankCode = Checks::checkParam($params, "cashierBankCode");
         $this-> reconciliationId = Checks::checkParam($params, "reconciliationId");
         $this-> tags = Checks::checkParam($params, "tags");
         $this-> uuid = Checks::checkParam($params, "uuid");